+++ CheckedExceptions.Tests/CheckedExceptionsAnalyzerTests.DeclaredSuperClassDetection.cs
@@ -24,7 +24,7 @@ partial class CheckedExceptionsAnalyzerTests
             }
             """;
 
-        var expected = Verifier.RedundantExceptionDeclaration("System.InvalidOperationException")
+        var expected = Verifier.RedundantExceptionDeclarationBySuperType("System.InvalidOperationException")
             .WithSpan(6, 13, 6, 44);
 
         await Verifier.VerifyAnalyzerAsync(test, expected);